# packages/schema/api.openapi.yaml
openapi: 3.1.0
info:
  title: Halyard API
  version: 0.1.0
  description: |
    The public REST surface, `api`. Paths and their semantics come from SPEC 7.1;
    resource shapes come from the SPEC 6 DDL.

    **Scope policy.** SPEC 7.1 lists every path in the product but specifies
    request and response bodies for none of them. This file therefore contains
    only the endpoints whose payloads the specification actually determines —
    the phase 0 and phase 1 surface. Later phases add their own paths in the
    task that implements them. Writing speculative bodies for the phase 6 ads
    endpoints now would be inventing an API surface, which SPEC 0 rule 5
    forbids, and would be churned rather than used.

    Endpoints specified in SPEC 7.1 and deliberately absent until their phase:
    tree/blob/file/diff and timeline/restore (phase 2), deployments and domains
    (phase 3), credits and the ledger (phase 4), capabilities, repo and
    proposals (phase 5), ads and seo (phase 6), and the webhook and relay
    receivers.

    **Conventions.**
    - Every mutating endpoint accepts `Idempotency-Key` (SPEC 7.1). Replaying a
      key returns the original response rather than acting twice, which is what
      makes a retry after a timeout safe (SPEC 17.2).
    - Errors use the envelope in `common.schema.json`: a stable `code` to branch
      on, a `message` saying what happened, and a `fix` saying what to do about
      it (SPEC 10, 18).
...
      inline (SPEC 8).
servers:
...
security:
  - sessionCookie: []
  - bearerToken: []
...
paths:
...
  /orgs:
    post:
      operationId: createOrg
      tags: [orgs]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrgRequest" }
      responses:
        "201":
          description: Created. The caller becomes its owner.
          content: { application/json: { schema: { $ref: "#/components/schemas/Org" } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
...
  /projects:
    post:
      operationId: createProject
      tags: [projects]
      description: |
        SPEC 7.1 takes either a template version or a prompt, never both: with a
        prompt the agent selects the template itself. SLO is p50 < 10s to first
        preview (SPEC 17.3), which is why this returns as soon as the project
        and branch rows exist rather than waiting for a sandbox.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProjectRequest" }
      responses:
        "201":
          description: Created.
          content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } }
        "402": { $ref: "#/components/responses/PaymentRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
  /projects/{project}:
    parameters: [{ $ref: "#/components/parameters/Project" }]
    get:
      operationId: getProject
      tags: [projects]
      responses:
        "200":
          description: The project and its branches.
          content: { application/json: { schema: { $ref: "#/components/schemas/Project" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteProject
      tags: [projects]
      description: |
        Requires `owner` (SPEC 8). Archives rather than erasing: SPEC 19.3 asks
        how long a suspended project's data is kept and how the user is warned
        before deletion, so erasure is a separate retention job.
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "204": { description: Archived. }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: halyard_session
      description: httpOnly, SameSite=Lax, rotating (SPEC 8).
    bearerToken:
      type: http
      scheme: bearer
      description: Personal access token, for the future CLI (SPEC 7.1).
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 255 }
      description: Replaying a key returns the original response instead of acting twice (SPEC 7.1, 17.2).
...
    Project:
      name: project
      in: path
      required: true
      x-go-name: ProjectSlugParam
      schema: { $ref: "./common.schema.json#/components/schemas/Slug" }
...
  responses:
    Invalid:
      description: The request was understood but is not valid.
      content:
        {
          application/json:
            { schema: { $ref: "./common.schema.json#/components/schemas/ErrorResponse" } },
        }
    Forbidden:
      description: Authenticated, but this role may not do this (SPEC 8).
      content:
        {
          application/json:
            { schema: { $ref: "./common.schema.json#/components/schemas/ErrorResponse" } },
        }
    NotFound:
      description: |
        No such resource, or none visible to this org. Cross-tenant reads are
        indistinguishable from absence on purpose.
      content:
        {
          application/json:
            { schema: { $ref: "./common.schema.json#/components/schemas/ErrorResponse" } },
        }
    Conflict:
      description: The resource's current state does not allow this.
      content:
        {
          application/json:
            { schema: { $ref: "./common.schema.json#/components/schemas/ErrorResponse" } },
        }
...
    PaymentRequired:
      description: |
        Insufficient credits to place a hold. SPEC 16.3: exhausting the build
        meter pauses the builder, which is harmless because the user is at the
        keyboard.
      content:
        {
          application/json:
            { schema: { $ref: "./common.schema.json#/components/schemas/ErrorResponse" } },
        }
  schemas:
...
    Org:
      type: object
      properties:
        id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
        name: { type: string, minLength: 1 }
        slug: { $ref: "./common.schema.json#/components/schemas/Slug" }
        plan: { type: string }
        created_at: { $ref: "./common.schema.json#/components/schemas/Timestamp" }
      required: [id, name, slug, plan, created_at]
      additionalProperties: false
    CreateOrgRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        slug: { $ref: "./common.schema.json#/components/schemas/Slug" }
      required: [name]
      additionalProperties: false
...
    CreateProjectRequest:
      description: Exactly one of `template_version_id` or `prompt` (SPEC 7.1).
      oneOf:
        - title: CreateProjectFromTemplate
          type: object
          properties:
            org_id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
            name: { type: string, minLength: 1, maxLength: 200 }
            template_version_id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
          required: [org_id, name, template_version_id]
          additionalProperties: false
        - title: CreateProjectFromPrompt
          type: object
          properties:
            org_id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
            name: { type: string, minLength: 1, maxLength: 200 }
            prompt: { type: string, minLength: 1, maxLength: 20000 }
          required: [org_id, prompt]
          additionalProperties: false
    Project:
      type: object
      properties:
        id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
        org_id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
        name: { type: string }
        slug: { $ref: "./common.schema.json#/components/schemas/Slug" }
        default_branch: { type: string }
        git_authority:
          $ref: "#/components/schemas/GitAuthority"
        branches: { type: array, items: { $ref: "#/components/schemas/Branch" } }
        archived_at: { $ref: "./common.schema.json#/components/schemas/Timestamp" }
        created_at: { $ref: "./common.schema.json#/components/schemas/Timestamp" }
      required: [id, org_id, name, slug, default_branch, git_authority, created_at]
      additionalProperties: false
    GitAuthority:
      title: GitAuthority
      enum: [internal, github]
      description: |
        SPEC 5.5: exactly one owner per project at all times. Once GitHub is
        linked it becomes authoritative and the internal store is a cache plus
        outbound queue. Dual authority on one branch is a split-brain bug
        factory; do not build it.
    Branch:
      type: object
      properties:
        id: { $ref: "./common.schema.json#/components/schemas/Uuid" }
        name: { type: string, minLength: 1 }
        head_sha: { $ref: "./common.schema.json#/components/schemas/GitSha" }
      required: [id, name]
      additionalProperties: false
...